oc-commands-analyzer-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["oc_commands_analyzer_agent"]

[project]
name = "oc-commands-analyzer-agent"
version = "0.1.0"
description = "AI agent for analyzing OpenShift CLI commands using MCP"
readme = "README.md"
keywords = ["agent", "openshift", "oc", "mcp", "langgraph"]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.9.0",
    "python-dotenv>=1.1.0",
    "structlog>=24.1.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-google-vertexai>=2.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.4.0",
    "langgraph-checkpoint>=2.0.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "sse-starlette>=2.3.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-asyncio==1.0.0",
    "pytest-cov==6.2.1",
    "ruff==0.12.2",
    "mypy==1.16.1",
]

[project.scripts]
oc-commands-analyzer-agent = "oc_commands_analyzer_agent.src.main:main"

[project.urls]
Homepage = "https://github.com/redhat-data-and-ai/oc-commands-analyzer-agent"
Repository = "https://github.com/redhat-data-and-ai/oc-commands-analyzer-agent"

[tool.ruff]
target-version = "py312"
line-length = 88
lint.select = ["E", "W", "F", "I", "F401"]
lint.ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = ["asyncio"]
